+++ vyper/evm/assembler/constants.py
